Break InputController to New Class
This breaks the input controller off to the InputController class. This stips more logic out of the core gameInstance class and should make it easier to do things like have an input thread in the future. This does currently leave the input queue logic in the gameInstance class. That would probably be good to migrate in the future as well.

Currently this adds no new functionality but it is the first step needed to add more advanced and robust input options
