+++ CMakeLists.txt
@@ -83,6 +83,7 @@ add_library(${PROJECT_NAME} SHARED
   src/main/engine/GfxController/src/DummyGfxController.cpp
   src/main/engine/GfxController/src/OpenGlGfxController.cpp
   src/main/engine/AnimationController/src/AnimationController.cpp
+  src/main/engine/Misc/src/InputController.cpp
   src/main/engine/Misc/src/DeltaTime.cpp
   src/main/engine/Misc/src/Image.cpp
   src/main/misc/src/config.cpp
